refactor: isolate ECC parameter runtime reports
